K11-669 Extract DrawCameraMappingArea from CameraImageComposite
- extracted DrawCameraMappingArea from CameraImageComposite
- extracted BeamCameraMapping process to BeamCameraMappingHelper
- moved RootCompositeAware to gda.giu.client to gda.client.event

+++ uk.ac.diamond.daq.client.gui.camera/src/uk/ac/diamond/daq/client/gui/camera/absorption/ROIStatisticRow.java
@@ -113,7 +113,7 @@ class ROIStatisticRow {
 		int[] step = new int[] { 1, 1 };
 
 		Dataset intDataset = DatasetUtils.convertToDataset(dataset.getSliceView(start, end, step));
-		this.value = (int) intDataset.sum(true);
+		this.value = (long) intDataset.sum(true);
 		this.valueLabel.setText(Long.toString(this.value));
 	}
 }
